Removing Min_datatransfer versions

+++ apps/c/mb_shsgc/Max_datatranser/MPI/checkop_kernel_seq_kernel.cpp
@@ -4,7 +4,7 @@
 
 //user function
 inline void checkop_kernel(const double *rho_new, const double *x, const double *rhoin, double *pre, double *post,
-  double *num) {
+  int *num) {
 
   double diff;
   diff = (rho_new[OPS_ACC0(0)] - rhoin[OPS_ACC2(0)]);
